+++ .env.example
@@ -1,1 +1,3 @@
 BOT_TOKEN=
+GUILD_ID=
+MOD_CHANNEL_ID=
