+++ src/config.ts
@@ -1,0 +1,14 @@
+import { Env } from '@(-.-)/env'
+import { config } from 'dotenv'
+import { z } from 'zod'
+
+config()
+
+export const EnvironmentSchema = z.object({
+  BOT_TOKEN: z.string(),
+  GUILD_ID: z.string(),
+  MOD_CHANNEL_ID: z.string(),
+})
+
+export const Environment = Env(EnvironmentSchema)
+export type Environment = z.infer<typeof EnvironmentSchema>
